+++ packages/standard_repositories/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1-dev.4
+
+- feat: make initial value optional
+- feat: make repositories not abstract
+
 ## 0.0.1-dev.3
 
 - feat: add multi repository